6-size.c

+++ 0x00-hello_world/6-size.c
@@ -10,11 +10,11 @@ int main(void)
 	long int c;
 	long long int d;
 	float f;
-    printf("Size of a char: %lu bytes(s)\n",(unsigned long) sizeof(a));
-       printf("Size of an int: %lu bytes(s)\n",(unsigned long) sizeof(b));
-       printf("Size of a long int: %lu bytes(s)\n",(unsigned long) sizeof(c));
-       printf("Size of a long long int: %lu bytes(s)\n",(unsigned long) sizeof(d));
-       printf("Size of a float: %lu bytes(s)\n",(unsigned long) sizeof(f));
+printf("Size of a char: %zu bytes(s)\n", sizeof(a));
+printf("Size of an int: %zu bytes(s)\n", sizeof(b));
+printf("Size of a long int: %zu bytes(s)\n", sizeof(c));
+printf("Size of a long long int: %zu bytes(s)\n", sizeof(d));
+printf("Size of a float: %zu bytes(s)\n", sizeof(f));
 
        return (0);
 }
